add flag to disable continuous context collection
